Remove comment which exception I was trying to hide
Problem persists: Either import slf4j simple in pom.xml and have
warnings like the comment show up OR do not import it and have a warning
about slf4j bindings.

We don't actually care about the content of the warning. We are not
validating the server fully. As we are trying to scan the server,
we would like to connect without full validation.